Fix Google Sheets integration
The changes for #1909 broke the Google Sheets integration by placing the spreadsheetId twice in the URL. Also the form-settings' spreadsheetId did not parse environment variables which broke implementations where an environment variable is set in the plugin-settings.
This fixes those issues. I was not able to change the spreadsheetId field in the form-settings for an autosuggest field, but environment variables are now being parsed.

I did not check the craft-4 version, that might also have these issues.